Move all reducers into reducers folder, remap imports
