+++ app/reducers/MovieCardReducers.js
